debug_26/4

+++ Proceduce_Function.sql
@@ -25,6 +25,68 @@ DROP TRIGGER IF EXISTS before_del_Bill//
 DROP TRIGGER IF EXISTS before_del_Billdetail//
 DROP TRIGGER IF EXISTS after_add_W_E//
 
+CREATE PROCEDURE add_w_e(
+    IN p_buildingid VARCHAR(50),
+    IN p_tenantid VARCHAR(50),
+    IN p_o_w VARCHAR(50),  -- Chỉ số nước cũ
+    IN p_n_w VARCHAR(50),  -- Chỉ số nước mới
+    IN p_o_e VARCHAR(50),  -- Chỉ số điện cũ
+    IN p_n_e VARCHAR(50),  -- Chỉ số điện mới
+    IN p_month VARCHAR(50) -- Tháng (1-12)
+)
+BEGIN
+   DECLARE new_ID VARCHAR(50);-- Id điện 
+   DECLARE new_ID_0 VARCHAR(50);-- Id nước 
+   
+   DECLARE v_start_date VARCHAR(50);-- ngày bắt đầu thanh toán
+   DECLARE v_end_date VARCHAR(50);-- ngày kết thúc thanh toán
+	DECLARE v_now VARCHAR(50);-- Ngày tạo hóa đơn
+	DECLARE new_id_2 VARCHAR(20);-- Id bill
+    
+   DECLARE v_total VARCHAR(20);  -- Tổng tiền add vào bill     
+   
+   DECLARE billdt_id_1 VARCHAR(20);
+   DECLARE billdt_id_2 VARCHAR(20);
+    
+    SET v_start_date = STR_TO_DATE(CONCAT('1/', p_month, '/', YEAR(CURDATE())), '%d/%m/%Y');
+    SET v_end_date = LAST_DAY(v_start_date);
+    SET v_now = Date(NOW());
+    
+    SET new_ID = createFigureID();
+    INSERT INTO WATER_ELECTRICITY (FIGUREID, UNITPRICEID, TENANTID, OLDFIGURE, NEWFIGURE, START_DATE, END_DATE,RECORD_DATE,TYPE,UNIT)
+    VALUES (new_ID, 'UP001',p_tenantID,p_o_e,  p_n_e, v_start_date,v_end_date,v_now,'ELECTRICITY','kWh');
+    
+    SET new_ID_0 = createFigureID();
+    INSERT INTO WATER_ELECTRICITY (FIGUREID, UNITPRICEID, TENANTID, OLDFIGURE, NEWFIGURE, START_DATE, END_DATE,RECORD_DATE,TYPE,UNIT)
+    VALUES (new_ID_0, 'UP002',  p_tenantID,  p_o_w, p_n_w, v_start_date, v_end_date,v_now,'WATER','m3');
+
+   SET new_id_2 = createBillID();
+
+	set billdt_id_1 = createBillDetailID();
+   INSERT INTO billdetail (BILLDETAIL_ID,BILLID, ID, AMOUNT)
+   SELECT billdt_id_1,new_id_2, us.SERVICEID, s.UNITPRICE 
+   FROM use_service us
+   JOIN service s ON s.SERVICEID = us.SERVICEID
+   WHERE us.TENANTID = p_tenantid
+   AND s.ISDELETED = 0
+   AND us.ISDELETED = 0;
+   
+   set billdt_id_2 = createBillDetailID();
+   INSERT INTO billdetail (BILLDETAIL_ID, BILLID, ID, AMOUNT)
+   SELECT billdt_id_2,new_id_2, we.FIGUREID,weu.UNITPRICE
+   FROM water_electricity we
+   JOIN WATER_ELECT_UNITPRICE weu ON weu.UNITPRICEID = we.UNITPRICEID
+   WHERE we.TENANTID = p_tenantid
+	AND (we.FIGUREID = new_ID_0 OR we.FIGUREID = new_ID)
+    AND we.ISDELETED = 0;
+   
+   SELECT SUM(AMOUNT) INTO v_total
+   FROM billdetail
+   WHERE BILLID = new_id_2;
+    
+   INSERT INTO bill (BILLID, TENANTID, TOTAL, START_DATE, END_DATE)
+   VALUES (new_id_2, p_tenantID, v_total, v_start_date, v_end_date);	
+END //
 CREATE PROCEDURE add_key(
     p_usern VARCHAR(20),
     p_key VARCHAR(20)
@@ -165,9 +227,93 @@ END//
 
 
 
-
+CREATE FUNCTION createBillDetailID()
+RETURNS VARCHAR(20) 
+DETERMINISTIC
+BEGIN
+    DECLARE max_id VARCHAR(20);
+    DECLARE number_part INT;
+    DECLARE new_id VARCHAR(20);
+    
+    SELECT IFNULL(MAX(BILLDETAIL_ID), 'BD000') INTO max_id FROM billdetail;
+    SET number_part = CAST(SUBSTRING(max_id, 3) AS UNSIGNED) + 1;
+    SET new_id = CONCAT('BD', LPAD(number_part, 3, '0'));
+    RETURN new_id;
+END//
 
 -- PAYMENT-BILL-W-E
+-- CREATE PROCEDURE load_billdetail(
+-- 	IN p_billID VARCHAR(20)
+-- )
+-- BEGIN
+--    SELECT bd.BILLDETAIL_ID,bd.BILLID,bd.ID,s.SERVICENAME,bd.AMOUNT FROM billdetail bd 
+--    JOIN service s ON s.SERVICEID = bd.ID
+--    WHERE bd.BILLID = p_billID AND bd.ISDELETED = 0 AND s.ISDELETED = 0
+-- 	union
+-- 	SELECT bd.BILLDETAIL_ID,bd.BILLID,bd.ID,
+-- 	CASE 
+--         WHEN w.`TYPE` = 'WATER' THEN 'Nước'
+--         WHEN w.`TYPE` = 'ELECTRICITY' THEN 'Điện'
+--         ELSE w.`TYPE`
+--    END,bd.AMOUNT FROM billdetail bd 
+--    JOIN water_electricity w ON w.FIGUREID = bd.ID
+-- 	WHERE bd.BILLID = p_billID AND bd.ISDELETED = 0 AND w.ISDELETED = 0;
+-- END//
+
+CREATE PROCEDURE calculate_bill()
+BEGIN
+	DECLARE bill_id VARCHAR(20);
+	
+	SELECT b.BILLID INTO bill_id FROM bill b
+	WHERE b.TOTAL IS NULL AND ISDELETED = 0
+	LIMIT 1;
+	
+   UPDATE bill 
+	SET bill.TOTAL = (
+		SELECT SUM(AMOUNT) FROM billdetail
+		WHERE billdetail.BILLID = bill_id AND ISDELETED = 0
+	),
+	bill.START_DATE = DATE_FORMAT(CURDATE(), '%Y-%m-01'),
+	bill.END_DATE = LAST_DAY(CURDATE())
+	WHERE bill.BILLID = bill_id;
+   
+END //
+
+CREATE PROCEDURE add_detailBill(
+    IN p_BILLID VARCHAR(50),
+    IN p_ID VARCHAR(50),
+   IN p_AMOUNT VARCHAR(50)
+)
+BEGIN
+    DECLARE v_billdetailid VARCHAR(50);
+    SET v_billdetailid = createBillDetailID();
+
+    INSERT INTO billdetail (
+       BILLDETAIL_ID,
+       BILLID,
+       ID,
+       AMOUNT
+    ) VALUES (
+      v_billdetailid,
+      p_BILLID,
+      p_ID,
+      p_AMOUNT
+    );
+
+END//
+
+CREATE PROCEDURE add_all_registration_Ser_into_billdetail(
+	IN p_billid VARCHAR(50),
+   IN p_ID VARCHAR(50),
+   IN p_AMOUNT VARCHAR(50)
+)
+BEGIN
+	DECLARE new_ID VARCHAR(50);
+	SET new_ID = createBillDetailID();
+	INSERT INTO billdetail (BILLDETAIL_ID,BILLID,ID,AMOUNT)
+	VALUES (new_ID,p_billid,p_ID,p_AMOUNT);
+END//
+
 CREATE PROCEDURE take_billid()
 BEGIN
    SELECT b.BILLID FROM bill b
@@ -207,6 +353,33 @@ BEGIN
 	LIMIT 100;
 END //
 
+-- CREATE PROCEDURE load_payment(
+-- 	IN p_buildingID VARCHAR(20),
+-- 	IN p_tenantid VARCHAR(20),
+-- 	IN p_lastname VARCHAR(50)
+-- )
+-- BEGIN
+-- 	DELETE FROM payment 
+--     WHERE ISDELETED = 1 AND DATEDIFF(CURDATE(), DELETED_DATE) > 30;
+
+--    SELECT p.PAYMENTID,t.TENANTID,t.FIRSTNAME,t.LASTNAME,p.BILLID,p.METHOD,p.TOTAL,Date(p.PAYMENTTIME)  FROM payment p 
+--    JOIN bill bi ON bi.BILLID = p.BILLID
+-- 	JOIN tenant t ON t.TENANTID = bi.TENANTID
+--    JOIN contract c ON c.TENANTID = t.TENANTID
+--    JOIN room r ON r.ROOMID = c.ROOMID
+--    JOIN building b ON b.BUILDINGID = r.BUILDINGID
+
+--    WHERE b.BUILDINGID = p_buildingID
+--     AND p.ISDELETED = 0
+--     AND bi.ISDELETED = 0
+--     AND t.ISDELETED = 0
+--     AND c.ISDELETED = 0
+--     AND r.ISDELETED = 0
+--     AND b.ISDELETED = 0
+-- 	AND (p_lastname IS NULL OR CONCAT(t.FIRSTNAME,' ',t.LASTNAME) LIKE CONCAT('%', p_lastname, '%'))
+-- 	AND (p_tenantid IS NULL OR t.TENANTID = p_tenantid);
+-- END//
+
 CREATE TRIGGER before_del_Bill
 BEFORE DELETE ON bill
  FOR EACH ROW
@@ -225,75 +398,43 @@ END//
 CREATE PROCEDURE del_bill(
     IN p_BillID VARCHAR(50)
 )
-BEGIN 	
-    UPDATE bill
-    SET ISDELETED = 1,
-        DELETED_DATE = CURDATE()
-    WHERE bill.BILLID = p_BillID;     
+BEGIN     
+    DECLARE var_id1 VARCHAR(50) DEFAULT NULL;
+    DECLARE var_id2 VARCHAR(50) DEFAULT NULL;
+    DECLARE record_count INT DEFAULT 0;
+    
+    -- Đếm số bản ghi trong billdetail
+    SELECT COUNT(*) INTO record_count 
+    FROM billdetail 
+    WHERE BILLID = p_BillID AND ISDELETED = 0;
+    
+    -- Chỉ thực hiện nếu có ít nhất 1 bản ghi
+    IF record_count > 0 THEN
+        -- Đánh dấu xóa trong billdetail
+        UPDATE billdetail
+        SET ISDELETED = 1,
+            DELETED_DATE = CURDATE()
+        WHERE BILLID = p_BillID AND ISDELETED = 0;  
+        
+        -- Đánh dấu xóa trong bill
+        UPDATE bill
+        SET ISDELETED = 1,
+            DELETED_DATE = CURDATE()
+        WHERE BILLID = p_BillID AND ISDELETED = 0;  
+        
+        -- Lấy 2 ID đầu tiên (nếu có)
+       
+        
+        -- Đánh dấu xóa trong water_electricity nếu có ID hợp lệ
+
+		UPDATE water_electricity 
+		SET ISDELETED = 1,
+			DELETED_DATE = CURDATE()
+		WHERE FIGUREID in (select ID from billdetail where BILLID = p_BillID)
+		AND ISDELETED = 0;
+    END IF;
 END//
 
-CREATE PROCEDURE add_w_e(
-    IN p_buildingid VARCHAR(50),
-    IN p_tenantid VARCHAR(50),
-    IN p_o_w VARCHAR(50),  -- Chỉ số nước cũ
-    IN p_n_w VARCHAR(50),  -- Chỉ số nước mới
-    IN p_o_e VARCHAR(50),  -- Chỉ số điện cũ
-    IN p_n_e VARCHAR(50),  -- Chỉ số điện mới
-    IN p_month VARCHAR(50) -- Tháng (1-12)
-)
-BEGIN
-   DECLARE new_ID VARCHAR(50);-- Id điện 
-   DECLARE new_ID_0 VARCHAR(50);-- Id nước 
-   
-   DECLARE v_start_date VARCHAR(50);-- ngày bắt đầu thanh toán
-   DECLARE v_end_date VARCHAR(50);-- ngày kết thúc thanh toán
-	DECLARE v_now VARCHAR(50);-- Ngày tạo hóa đơn
-	DECLARE new_id_2 VARCHAR(20);-- Id bill
-    
-   DECLARE v_total VARCHAR(20);  -- Tổng tiền add vào bill     
-   
-   DECLARE billdt_id_1 VARCHAR(20);
-   DECLARE billdt_id_2 VARCHAR(20);
-    
-    SET v_start_date = STR_TO_DATE(CONCAT('1/', p_month, '/', YEAR(CURDATE())), '%d/%m/%Y');
-    SET v_end_date = LAST_DAY(v_start_date);
-    SET v_now = Date(NOW());
-    
-    SET new_ID = createFigureID();
-    INSERT INTO WATER_ELECTRICITY (FIGUREID, UNITPRICEID, TENANTID, OLDFIGURE, NEWFIGURE, START_DATE, END_DATE,RECORD_DATE,TYPE,UNIT)
-    VALUES (new_ID, 'UP001',p_tenantID,p_o_e,  p_n_e, v_start_date,v_end_date,v_now,'ELECTRICITY','kWh');
-    
-    SET new_ID_0 = createFigureID();
-    INSERT INTO WATER_ELECTRICITY (FIGUREID, UNITPRICEID, TENANTID, OLDFIGURE, NEWFIGURE, START_DATE, END_DATE,RECORD_DATE,TYPE,UNIT)
-    VALUES (new_ID_0, 'UP002',  p_tenantID,  p_o_w, p_n_w, v_start_date, v_end_date,v_now,'WATER','m3');
-
-   SET new_id_2 = createBillID();
-
-	set billdt_id_1 = createBillDetailID();
-   INSERT INTO billdetail (BILLDETAIL_ID,BILLID, ID, AMOUNT)
-   SELECT billdt_id_1,new_id_2, us.SERVICEID, s.UNITPRICE 
-   FROM use_service us
-   JOIN service s ON s.SERVICEID = us.SERVICEID
-   WHERE us.TENANTID = p_tenantid
-   AND s.ISDELETED = 0
-   AND us.ISDELETED = 0;
-   
-   set billdt_id_2 = createBillDetailID();
-   INSERT INTO billdetail (BILLDETAIL_ID, BILLID, ID, AMOUNT)
-   SELECT billdt_id_2,new_id_2, we.FIGUREID,weu.UNITPRICE
-   FROM water_electricity we
-   JOIN WATER_ELECT_UNITPRICE weu ON weu.UNITPRICEID = we.UNITPRICEID
-   WHERE we.TENANTID = p_tenantid
-	AND (we.FIGUREID = new_ID_0 OR we.FIGUREID = new_ID)
-    AND we.ISDELETED = 0;
-   
-   SELECT SUM(AMOUNT) INTO v_total
-   FROM billdetail
-   WHERE BILLID = new_id_2;
-    
-   INSERT INTO bill (BILLID, TENANTID, TOTAL, START_DATE, END_DATE)
-   VALUES (new_id_2, p_tenantID, v_total, v_start_date, v_end_date);	
-END //
 
 CREATE PROCEDURE del_W_E(
     IN p_tenantID VARCHAR(50),
@@ -324,26 +465,28 @@ BEGIN
     DELETE FROM water_electricity 
     WHERE ISDELETED = 1 AND DATEDIFF(CURDATE(), DELETED_DATE) > 30;
 
-    SELECT we.FIGUREID, we.UNITPRICEID
-, we.TENANTID 
-, we.OLDFIGURE  
-, we.NEWFIGURE  
-, we.UNIT 
-, we.START_DATE  
-, we.END_DATE  
-, we.RECORD_DATE  
-, we.TYPE
- FROM water_electricity we
+SELECT 
+    we.FIGUREID,
+    we.UNITPRICEID,
+    we.TENANTID,
+    we.OLDFIGURE,
+    we.NEWFIGURE,
+    we.UNIT,
+    we.START_DATE,
+    we.END_DATE,
+    we.RECORD_DATE,
+    we.TYPE
+FROM water_electricity we
 	JOIN tenant t ON t.TENANTID = we.TENANTID
 	JOIN contract c ON c.TENANTID = t.TENANTID
-	JOIN room r ON r.ROOMID = c.ROOMID
-	JOIN building b ON b.BUILDINGID = r.BUILDINGID
-	WHERE b.BUILDINGID = p_buildingid
-	AND t.ISDELETED = 0
-    AND we.ISDELETED = 0
-    AND c.ISDELETED = 0
-    AND room.ISDELETED = 0
-	AND t.USERNAME =  p_user;
+    JOIN room r ON r.ROOMID = c.ROOMID 
+    JOIN building b ON b.BUILDINGID = r.BUILDINGID
+WHERE b.BUILDINGID = p_buildingid
+        AND t.ISDELETED = 0
+        AND we.ISDELETED = 0
+        AND c.ISDELETED = 0
+        AND r.ISDELETED = 0
+        AND t.USERNAME = p_user;
 END //
 
 CREATE FUNCTION createFigureID()
@@ -399,7 +542,9 @@ BEGIN
 	JOIN contract c ON c.TENANTID = t.TENANTID
 	Join room r on r.ROOMID = c.ROOMID
 	WHERE r.ROOMNAME = p_room_id
-    AND t.ISDELETED = 0;
+    AND t.ISDELETED = 0
+    AND c.ISDELETED = 0
+    AND r.ISDELETED = 0;
 END //
 
 CREATE PROCEDURE load_bill(
@@ -1413,37 +1558,55 @@ BEGIN
     -- Delete records that have been soft-deleted for more than 30 days
     DELETE FROM ROOM 
     WHERE ISDELETED = 1 AND DATEDIFF(CURDATE(), DELETED_DATE) > 30;
+    
+    IF p_buildingid is null then
+		SELECT 
+			r.ROOMNAME, 
+			r.BUILDINGID, 
+			r.TYPE, 
+            r.FLOOR,
+			r.CONVENIENT, 
+			r.AREA, 
+			r.PRICE, 
+			r.STATUS
+		FROM ROOM r
+		JOIN BUILDING b ON r.BUILDINGID = b.BUILDINGID
+		JOIN USER u ON u.USERNAME = b.USERNAME
+		WHERE u.USERNAME = p_username
+		AND r.ISDELETED = 0;
+	ELSE
 
-    -- Tạo bảng tạm lưu trạng thái
-    CREATE TEMPORARY TABLE IF NOT EXISTS temp_statuses (
-        status_value VARCHAR(100)
-    );
-    TRUNCATE TABLE temp_statuses;
-    
-    -- Chèn các trạng thái vào bảng tạm
-    SET @sql = CONCAT('INSERT INTO temp_statuses VALUES ("', 
-                     REPLACE(p_status_list, '; ', '"),("'), '")');
-    PREPARE stmt FROM @sql;
-    EXECUTE stmt;
-    DEALLOCATE PREPARE stmt;
-    
-    -- Đếm tổng số trạng thái cần kiểm tra
-    SET @total_statuses = (SELECT COUNT(*) FROM temp_statuses);
-    
-    -- Lọc phòng phải chứa TẤT CẢ trạng thái được chọn
-    SELECT r.ROOMNAME, r.BUILDINGID, r.TYPE, r.FLOOR, r.CONVENIENT, r.AREA, r.PRICE, r.STATUS
-    FROM ROOM r
-    JOIN BUILDING b ON r.BUILDINGID = b.BUILDINGID
-    WHERE b.USERNAME = p_username 
-    AND r.buildingid = p_buildingid
-    AND r.ISDELETED = 0
-    AND (
-        SELECT COUNT(*) 
-        FROM temp_statuses ts 
-        WHERE r.STATUS LIKE CONCAT('%', ts.status_value, '%')
-    ) = @total_statuses;
-    
-    DROP TEMPORARY TABLE IF EXISTS temp_statuses;
+		-- Tạo bảng tạm lưu trạng thái
+		CREATE TEMPORARY TABLE IF NOT EXISTS temp_statuses (
+			status_value VARCHAR(100)
+		);
+		TRUNCATE TABLE temp_statuses;
+		
+		-- Chèn các trạng thái vào bảng tạm
+		SET @sql = CONCAT('INSERT INTO temp_statuses VALUES ("', 
+						 REPLACE(p_status_list, '; ', '"),("'), '")');
+		PREPARE stmt FROM @sql;
+		EXECUTE stmt;
+		DEALLOCATE PREPARE stmt;
+		
+		-- Đếm tổng số trạng thái cần kiểm tra
+		SET @total_statuses = (SELECT COUNT(*) FROM temp_statuses);
+		
+		-- Lọc phòng phải chứa TẤT CẢ trạng thái được chọn
+		SELECT r.ROOMNAME, r.BUILDINGID, r.TYPE, r.FLOOR, r.CONVENIENT, r.AREA, r.PRICE, r.STATUS
+		FROM ROOM r
+		JOIN BUILDING b ON r.BUILDINGID = b.BUILDINGID
+		WHERE b.USERNAME = p_username 
+		AND r.buildingid = p_buildingid
+		AND r.ISDELETED = 0
+		AND (
+			SELECT COUNT(*) 
+			FROM temp_statuses ts 
+			WHERE r.STATUS LIKE CONCAT('%', ts.status_value, '%')
+		) = @total_statuses;
+		
+		DROP TEMPORARY TABLE IF EXISTS temp_statuses;
+	end if;
 END//
 
  CREATE DEFINER=`root`@`localhost` FUNCTION IF NOT EXISTS `checkFloorCapacity`(
@@ -2060,6 +2223,15 @@ END//
 
 
 -- SERVICE
+CREATE PROCEDURE load_registration_service_to_add(
+	IN p_tenant_id VARCHAR(20)
+)
+BEGIN
+   SELECT us.SERVICEID,s.UNITPRICE FROM use_service us
+   JOIN service s ON s.SERVICEID = us.SERVICEID
+   WHERE us.TENANTID = p_tenant_id AND us.ISDELETED = 0 AND s.ISDELETED = 0;
+END //
+
 CREATE PROCEDURE INSERT_SERVICE_USAGE(
     IN P_TENANTID VARCHAR(10),
     IN P_SERVICEID VARCHAR(10),
